Updating to 2712

+++ src/net/http/http_security_headers.h
@@ -18,11 +18,11 @@ class GURL;
 
 namespace net {
 
-const int64_t kMaxHSTSAgeSecs = 86400 * 365;  // 1 year
+const uint32_t kMaxHSTSAgeSecs = 86400 * 365;  // 1 year
 
 // RFC7469 suggests that 60 days is a reasonable maximum max-age value
 // http://tools.ietf.org/html/rfc7469#section-4.1
-const int64_t kMaxHPKPAgeSecs = 86400 * 60;  // 60 days
+const uint32_t kMaxHPKPAgeSecs = 86400 * 60;  // 60 days
 
 // Parses |value| as a Strict-Transport-Security header value. If successful,
 // returns true and sets |*max_age| and |*include_subdomains|.
